Preload and cache AVPlayerItems
The L5PlayerPreloader class preloads AVPlayerItems by associating them
with an AVPlayer. In order to make this work more or less seamlessly, the
L5MultiPlayer class is used to manage all of the AVPlayer instances. The one
hack is that ShowViewController needs to update the AVPlayerLayer whenever
the video changes.