Fix bug where UI gets blocked if video not loaded
Previously, the code would only add items to the queue if the asset
was ready to play. However, this meant when the next item wasn't yet
preloaded, there was no indication to the user. Now, all items are
queued to the MultiPlayer immediately and if the current item to be
played isn't yet ready, a loading indicator is displayed. When the
item finishes loading, the indicator is hidden and the video starts
playing.